refactor(e2e): fold e2e_cheat_codes into @aztec/ethereum tests
src/e2e_cheat_codes.test.ts exercises EthCheatCodes directly against a raw
anvil with no Aztec node. EthCheatCodes and startAnvil are defined in
@aztec/ethereum (only re-exported by @aztec/aztec/testing), so the test
belongs there. Merge its mine/timestamp/load-store/etch/impersonate cases
into the existing ethereum/src/test/eth_cheat_codes.test.ts suite, reusing
its shared anvil setup. No new dependencies; all imports stay intra-package
or downward to @aztec/foundation.

Part of A-1176
